Update:
Somewhere the code became wonky and we had to debug.  Ive fixed the code so that it would no longer throw the errors when migration process was run...

			Warning: 1265 Data truncated for column

The error we received was the warning listed above. In order to combat this we changed the INT assignment to a DECIMAL assignment.

+++ database/migrations/user_migration.php
@@ -26,7 +26,7 @@ $query = 'CREATE TABLE listing (
     id INT UNSIGNED NOT NULL AUTO_INCREMENT,
     item_name VARCHAR(240) NOT NULL,
     item_description VARCHAR(500) NOT NULL,
-    item_price INT(100) NOT NULL,
+    item_price DECIMAL(65) NOT NULL,
     item_location_city VARCHAR(255) NOT NULL,
     item_location_state VARCHAR(2) NOT NULL,
     user_id INT UNSIGNED NOT NULL,
